requirements.txt
absl-py==1.4.0
addict==2.4.0
affine==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
ansi2html==1.9.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==22.2.0
Automat==22.10.0
av==10.0.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
bitarray==2.8.5
bleach==6.0.0
blinker==1.6.3
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1677499911949/work
brotlipy==0.7.0
bs4==0.0.1
build==1.0.3
cachetools==5.3.1
capstone==5.0.0.post1
Cartopy==0.21.1
cdsapi==0.6.1
certifi==2023.5.7
cffi==1.15.1
cftime==1.6.3
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cmake==3.26.4
cmdstanpy==1.2.0
colored-traceback==0.3.0
comm==0.1.2
conda @ file:///home/conda/feedstock_root/build_artifacts/conda_1694556045812/work
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
ConfigArgParse==1.7
constantly==15.1.0
contourpy==1.0.7
cramjam==2.6.2
cryptography @ file:///croot/cryptography_1673298753778/work
cssselect==1.2.0
cycler==0.11.0
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.9.0
Django==4.2
djitellopy==2.4.0
dnspython==2.4.2
docopt==0.6.2
docx==0.2.4
dotenv-python==0.0.1
exceptiongroup==1.1.1
executing==1.2.0
fake-useragent==1.1.3
fastjsonschema==2.16.3
fastparquet==2023.7.0
ffmpeg-python==0.2.0
filelock==3.12.2
Fiona==1.9.2
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gast==0.4.0
GDAL @ file:///croot/gdal-split_1678138042368/work/build/swig/python
geographiclib==2.0
geopandas==0.12.2
geopy==2.3.0
georaster==1.25
Glances==3.4.0.3
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
holidays==0.35
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.17.3
hyperlink==21.0.0
idna @ file:///croot/idna_1666125576474/work
image==1.5.33
importlib-metadata==6.7.0
importlib-resources==6.1.0
incremental==22.10.0
intervaltree==3.1.0
ipykernel==6.21.2
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jax==0.4.11
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.2
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.12.0
kiwisolver==1.4.4
langchain==0.0.320
langsmith==0.0.49
Levenshtein==0.23.0
libclang==16.0.0
lit==16.0.6
loguru==0.5.3
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
munch==2.5.0
mypy-extensions==1.0.0
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
NearPy==1.0.0
nest-asyncio==1.5.6
netCDF4==1.6.5
networkx==3.1
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
open3d==0.18.0
openai==1.11.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
osqp==0.6.3
outcome==1.2.0
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.0
paramiko==3.2.0
parse==1.19.1
parsel==1.8.1
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client @ git+https://git@github.com/pinecone-io/pinecone-python-client.git@591b8fb47a9a4fad7b102ce637733217796c452c
pip-tools==7.3.0
pipenv==2023.3.20
pipreqs==0.4.13
platformdirs==3.0.0
plotly==5.18.0
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
plumbum==1.8.2
plyfile==0.8.1
podaac-data-subscriber==1.14.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
prophet==1.1.5
Protego==0.2.1
protobuf==4.23.2
psutil==5.9.4
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pwntools==4.10.0
py4j==0.10.9.7
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
PyDispatcher==2.0.7
pydub==0.25.1
pyee==8.2.2
pyelftools==0.29
Pygments==2.14.0
pymongo==4.6.0
Pympler==1.0.1
PyNaCl==1.5.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pypdf==3.17.2
PyPDF2==3.0.1
pyppeteer==1.0.2
pyproj==3.5.0
pyproject_hooks==1.0.0
pyquaternion==0.9.9
pyquery==2.0.0
pyrealsense2==2.53.1.4623
pyrsistent==0.19.3
pyserial==3.5
pyshp==2.3.1
PySocks==1.7.1
pyspark==3.4.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-json-logger==2.0.7
pytorch-model-summary==0.1.2
pytrends==4.9.2
pytz==2023.3
pyvips==2.2.1
PyYAML==6.0.1
pyzmq==25.0.0
qdldl==0.1.7.post0
qtconsole==5.4.0
QtPy==2.3.0
queuelib==1.6.2
rapidfuzz==3.5.2
rasterio==1.3.6
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-html==0.10.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ROPGadget==7.3
rpyc==5.3.1
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
safetensors==0.4.0
scikit-learn==1.2.2
scipy==1.10.1
scrapingbee==1.2.0
Scrapy==2.9.0
selenium==4.10.0
selenium-requests==2.0.3
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==23.1.0
shapely==2.0.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post4
sniffio==1.3.0
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@614d4c2029a62d348ca56598f87c425966aaec66
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.17
sqlparse==0.4.4
stack-data==0.6.2
stanio==0.3.0
statsmodels==0.14.0
sympy==1.12
tellopy @ file:///home/dylaneg/Documents/Programming/IROM/tello-zoedepth-new/TelloPy/dist/tellopy-0.7.0.dev0-py2.py3-none-any.whl#sha256=9551e4fc6b8fbcba6e2fff247b156fb56dc8f0a1eed7fb8929ad38082f962614
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tiktoken==0.5.1
timm==0.6.7
tinycss2==1.2.1
tldextract==3.4.4
tokenizers==0.14.1
tomli==2.0.1
toolz @ file:///croot/toolz_1667464077321/work
torch==2.0.1
torch-tb-profiler==0.4.1
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.35.0
trio==0.22.0
trio-websocket==0.10.3
triton==2.0.0
Twisted==22.10.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.8.0
unicorn==2.0.1.post1
uri-template==1.2.0
urllib3 @ file:///croot/urllib3_1673575502006/work
vaderSentiment==3.3.2
virtualenv==20.21.0
w3lib==2.1.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==3.0.0
widgetsnbextension==4.0.5
wrapt==1.14.1
wsproto==1.2.0
xarray==2023.5.0
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
zstandard @ file:///opt/conda/conda-bld/zstandard_1663827383994/work
